backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.0.4
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
Crawl4AI==0.5.0.post8
cryptography==44.0.2
cssselect==1.3.0
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
durationpy==0.9
ecdsa==0.19.1
email_validator==2.2.0
emoji==2.14.1
eval_type_backport==0.2.2
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.115.9
fastapi-cli==0.0.7
faust-cchardet==2.1.19
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
humanize==4.12.2
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.31
litellm==1.66.0
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
oauthlib==3.2.2
olefile==0.47
ollama==0.4.7
onnxruntime==1.21.0
openai==1.74.0
opentelemetry-api==1.32.0
opentelemetry-exporter-otlp-proto-common==1.32.0
opentelemetry-exporter-otlp-proto-grpc==1.32.0
opentelemetry-instrumentation==0.53b0
opentelemetry-instrumentation-asgi==0.53b0
opentelemetry-instrumentation-fastapi==0.53b0
opentelemetry-proto==1.32.0
opentelemetry-sdk==1.32.0
opentelemetry-semantic-conventions==0.53b0
opentelemetry-util-http==0.53b0
orjson==3.10.16
overrides==7.7.0
packaging==24.2
passlib==1.7.4
pillow==10.4.0
playwright==1.51.0
posthog==3.24.1
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyee==12.1.1
Pygments==2.19.1
pyOpenSSL==25.0.0
pypdf==5.4.0
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.45.3
sympy==1.13.1
tenacity==9.1.2
tf-playwright-stealth==1.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
ujson==5.10.0
unstructured==0.17.2
unstructured-client==0.32.3
urllib3==2.4.0
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
